CorraiRAG/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
attrs==24.3.0
backoff==2.2.1
certifi==2024.12.14
charset-normalizer==3.4.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
dateparser==1.2.0
decorator==5.1.1
dill==0.3.9
distro==1.9.0
executing==2.1.0
faiss-cpu==1.9.0.post1
fonttools==4.55.3
frozenlist==1.5.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
imageio==2.36.1
ipython==8.31.0
ipywidgets==8.1.5
jedi==0.19.2
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-openai==0.3.2
langchain-text-splitters==0.3.5
langsmith==0.3.1
llvmlite==0.43.0
marshmallow==3.26.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mem0ai==0.1.48
monotonic==1.6
multidict==6.1.0
mypy-extensions==1.0.0
numba==0.60.0
numpy==2.0.2
openai==1.60.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
plotly==5.24.1
portalocker==2.10.1
posthog==3.9.3
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.3
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.13.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
stack-data==0.6.3
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
vectorbt==0.27.1
wcwidth==0.2.13
wheel==0.44.0
widgetsnbextension==4.0.13
yarl==1.18.3
zstandard==0.23.0
rank_bm25==0.2.0